Add for loop methods on Cursor
Added special methods to allow using for loops
on `Cursor` objects. Using a for loop on a cursor
after an executed query with a result set will
fetch rows at every iteration, until there is none
left.

Using that loop when there is no result set in
the cursor will result in the default exception
raised by `fetch` methods in the same situation.